Moved the services that could be moved from the `engine` sub-project to a new sub-project `services`.
