+++ build.sbt
@@ -16,6 +16,14 @@ lazy val database = (project in file("database"))
   .configs(CromwellNoIntegrationTest).settings(inConfig(CromwellNoIntegrationTest)(Defaults.testTasks): _*)
   .configs(DbmsTest).settings(inConfig(DbmsTest)(Defaults.testTasks): _*)
 
+lazy val services = (project in file("services"))
+  .dependsOn(core % "test->test;compile->compile")
+  .dependsOn(database % "test->test;compile->compile") // Assuming services impl in services
+  .settings(servicesSettings:_*)
+  .configs(AllTests).settings(inConfig(AllTests)(Defaults.testTasks): _*)
+  .configs(CromwellIntegrationTest).settings(inConfig(CromwellIntegrationTest)(Defaults.testTasks): _*)
+  .configs(CromwellNoIntegrationTest).settings(inConfig(CromwellNoIntegrationTest)(Defaults.testTasks): _*)
+
 lazy val backendRoot = Path("supportedBackends")
 
 lazy val backend = (project in file("backend"))
@@ -41,6 +49,7 @@ lazy val jesBackend = (project in backendRoot / "jes")
 //TODO: remove jesBackend once refactoring has finished.
 lazy val engine = (project in file("engine"))
   .dependsOn(core % "test->test;compile->compile",
+    services % "test->test;compile->compile",
     backend % "test->test;compile->compile",
     jesBackend % "test->test;compile->compile",
     localBackend % "test->test;compile->compile",
