+++ services/src/main/scala/cromwell/services/ServiceRegistryClient.scala
@@ -1,9 +1,10 @@
 package cromwell.services
 
 import com.typesafe.config.ConfigFactory
-import cromwell.server.WorkflowManagerSystem
 
-private object ServiceRegistryInstance extends WorkflowManagerSystem {
+private object ServiceRegistryInstance {
+  // TODO: PBE: Removed engine's WMS trait creating a singleton actor system... can't we just pass actor refs?
+  private val actorSystem = akka.actor.ActorSystem("cromwell-service-registry-system")
   val ServiceRegistryActorInstance = actorSystem.actorOf(ServiceRegistryActor.props(ConfigFactory.load()), "ServiceRegistryActor")
 }
